Add basic auth with PostgreSQL
